rename NewTableEntryDialog.java to NewContactDialog.java
